refactor: clarify cache module API surfaces, address review feedback
Restructure the cache module so the two audiences are clear:
- mod.rs: user-facing API (LanceCache, CacheKey, UnsizedCacheKey)
- backend.rs: implementor-facing API (CacheBackend, InternalCacheKey)

Delete keys.rs, moving CacheKey/UnsizedCacheKey into mod.rs and
InternalCacheKey into backend.rs.

Also: rename has_prefix -> starts_with, trim verbose get_or_insert doc,
add shared-buffer note to approx_size_bytes, improve type_name docs,
explain std::any::type_name usage in previous/reader.rs.
